# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bihar Cyber Café Digital Service Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [Users]
      summary: Get or JIT-provision the current user's profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateMe
      tags: [Users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: List users (admin only)
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UserRole"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [Users]
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /tickets:
    get:
      operationId: listTickets
      tags: [Tickets]
      summary: List support tickets (own for customers, all for operator/admin)
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createTicket
      tags: [Tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /tickets/{id}:
    patch:
      operationId: updateTicket
      tags: [Tickets]
      summary: Update a ticket's status/reply (operator/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserRole:
      type: string
      enum: [customer, operator, admin]
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
...
    TicketStatus:
      type: string
      enum: [open, in_progress, resolved]
    Ticket:
      type: object
      required:
        [
          id,
          userId,
          subject,
          message,
          status,
          adminReply,
          applicationId,
          createdAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        adminReply:
          type: ["string", "null"]
        applicationId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    TicketInput:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        applicationId:
          type: integer
    TicketUpdate:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/TicketStatus"
        adminReply:
          type: string
...