# lib/api-zod/src/generated/types/ticketStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Cyber Café Digital Service Portal API
 * OpenAPI spec version: 0.1.0
 */

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
} as const;