# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Cyber Café Digital Service Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  createdAt: Date;
}